+++ module/MonsterMacros.js
@@ -1,4 +1,4 @@
-class HomebrewMacros {
+class MonsterMacros {
 
     static async ghastStenchEffect(token) {
         logger.info("ghastStenchEffect");