+++ scripts/spells/heat-metal.js
@@ -10,9 +10,9 @@
 
 	At Higher Levels. When you cast this spell using a spell slot of 3rd level or higher, the damage increases by 1d8 for each slot level above 2nd.
 */
-const version = "11.0";
+const version = "11.1";
 const optionName = "Heat Metal";
-const metalWeapons = ["Scimitar", "sword", "blade", "Rapier", "dagger", "mace"];
+const metalWeapons = ["scimitar", "sword", "blade", "rapier", "dagger", "mace", "knife"];
 const metalArmor = ["medium", "heavy"];
 const gameRound = game.combat ? game.combat.round : 0;
 const flagName = "heat-metal";
@@ -21,9 +21,9 @@ const mutationName = "heat-metal-mutation";
 
 try {
 	if (args[0].macroPass === "postActiveEffects") {
-		const spellLevel = workflow.castData.castLevel;
-		const damageDice = 2 + Math.max(spellLevel - 2, 0);
+		const damageDice = workflow.castData.castLevel;
 		const target = workflow.targets.first();
+		let targetItemName;
 
 		// find the target actor's weapons & armor that are metal
 		let firstWeapons = target.actor.items.filter(i => (i.type === `weapon` && i.system.equipped));
@@ -32,7 +32,7 @@ try {
 		
 		for (let weapon of firstWeapons) {
 			for (var i = 0; i < metalWeapons.length; i++) {
-				if (weapon.name.includes(metalWeapons[i])) {
+				if (weapon.name.toLowerCase().includes(metalWeapons[i])) {
 					targetItems.add(weapon);
 					break;
 				}
@@ -71,10 +71,10 @@ try {
 						callback: async (html) => {
 							let itemId = html.find('[name=titem]')[0].value;
 							let selectedItem = target.actor.items.get(itemId);
-							const itemName = selectedItem.name;
+							targetItemName = selectedItem.name;
 
 							// apply the heat
-							const newName = itemName + ` (${optionName})`;
+							const newName = targetItemName + ` (${optionName})`;
 							
 							let mutations = {};
 							mutations[selectedItem.name] = {
@@ -88,11 +88,10 @@ try {
 							};
 
 							// mutate the selected item
-							await warpgate.mutate(target.document, updates, {}, { name: itemName });
+							await warpgate.mutate(target.document, updates, {}, { name: targetItemName });
 							// TODO add in option to drop the item, for now handled manually
 
-
-							// Apply the actor effect
+							// Apply the heated effect to the target
 							let actorEffectData = {
 								'name': effectName,
 								'icon': workflow.item.img,
@@ -124,11 +123,7 @@ try {
 								}
 							};
 							await MidiQOL.socket().executeAsGM('createEffects', {'actorUuid': target.actor.uuid, 'effects': [actorEffectData]});
-
-
-							// track target info on the source actor
-							DAE.setFlag(actor, flagName, {tokenId: target.id, itemName: itemName, origin: workflow.item.uuid, damageDice: damageDice});
-							ChatMessage.create({content: target.actor.name + "'s " + itemName + " becomes red-hot!"});
+							ChatMessage.create({content: target.actor.name + "'s " + targetItemName + " becomes red-hot!"});
 							resolve(true);
 						}
 					}
@@ -138,64 +133,34 @@ try {
 		
 		let useFeature = await dialog;
 		if (useFeature) {
+			// track target info to the caster
+			await actor.setFlag("fvtt-trazzm-homebrew-5e", flagName, {tokenId: target.id, origin: actor.uuid, itemName: targetItemName});
+
+			// add the ongoing damage item to the caster
+			let ongoingDamageItem = await HomebrewHelpers.getItemFromCompendium('fvtt-trazzm-homebrew-5e.homebrew-automation-items', 'Heat Metal Damage');
+			if (ongoingDamageItem) {
+				let damageParts = ongoingDamageItem.system.damage.parts;
+				damageParts[0][0] = `${damageDice}${damageParts[0][0].substring(1)}`;
+				ongoingDamageItem.system.damage.parts = damageParts;
+
+				const updates = {
+					embedded: { Item: { ['Heat Metal Damage']: ongoingDamageItem } }
+				}
+				await warpgate.mutate(token.document, updates, {}, {name: mutationName});
+			}
+
+			// Apply the initial damage to the target
 			const actorData = actor.getRollData();
 			let damageRoll = await new game.dnd5e.dice.DamageRoll(`${damageDice}d8[fire]`, actorData).evaluate({async:false});
-			await new MidiQOL.DamageOnlyWorkflow(actor, token, damageRoll.total, "fire", [target], damageRoll, { flavor: `(${optionName})`, itemCardId: "new" });
 			await game.dice3d?.showForRoll(damageRoll);
-		}
-		return useFeature;		
-	}
-	else if (args[0] === "each") {
-		// ask if they want to apply the damage as a bonus action
-		let effectedTarget = undefined;
-		let flag = DAE.getFlag(actor, flagName);
-		if (flag) {
-			// get the target actor
-			const targetToken = canvas.scene.tokens.get(flag.tokenId);
-			if (targetToken) {
-				let effect = targetToken.actor.effects.find(f => f.name === effectName && f.origin === flag.origin);
-				if (effect) {
-					effectedTarget = targetToken;
-				}
-			}
-		}
-
-		if (effectedTarget) {
-			let dialog = new Promise((resolve, reject) => {
-				new Dialog({
-					title: optionName,
-					content: `<p>Use a bonus action to apply damage to ${effectedTarget.name}?</p>`,
-					buttons: {
-						yes: {
-							icon: '<p> </p><img src = "icons/tools/smithing/furnace-fire-metal-orange.webp" width="50" height="50"></>',
-							label: "<p>Yes</p>",
-							callback: () => resolve(true)
-						},
-						no: {
-							icon: '<p> </p><img src = "icons/skills/melee/weapons-crossed-swords-yellow.webp" width="50" height="50"></>',
-							label: "<p>No</p>",
-							callback: () => {
-								resolve(false)
-							}
-						}
-					},
-					default: "two"
-				}).render(true);
-			});
-
-			let useFeature = await dialog;
-			if (useFeature) {
-				// apply damage to effectedTarget
-				const actorData = actor.getRollData();
-				let damageRoll = await new game.dnd5e.dice.DamageRoll(`${flag.damageDice}d8[fire]`, actorData).evaluate({async:false});
-				await new MidiQOL.DamageOnlyWorkflow(actor, token, damageRoll.total, "fire", [effectedTarget], damageRoll, { flavor: `(${optionName})`, itemCardId: "new" });
-				await game.dice3d?.showForRoll(damageRoll);
-			}
+			await new MidiQOL.DamageOnlyWorkflow(actor, token, damageRoll.total, "fire", [target], damageRoll, { flavor: `(${optionName})`, itemCardId: args[0].itemCardId });
 		}
 	}
 	else if (args[0] === "off") {
-		let flag = DAE.getFlag(actor, flagName);
+		let flag = actor.getFlag("fvtt-trazzm-homebrew-5e", flagName);
 		if (flag) {
+			await actor.unsetFlag("fvtt-trazzm-homebrew-5e", flagName);
+
 			// get the target actor
 			let targetToken = canvas.scene.tokens.get(flag.tokenId);
 			if (targetToken) {
@@ -204,7 +169,7 @@ try {
 				await warpgate.revert(targetToken, itemName);
 
 				// remove the effect
-				let effect = targetToken.actor.effects.find(f => f.name === effectName);
+				let effect = targetToken.actor.effects.find(f => f.name === effectName && f.origin === flag.origin);
 				if (effect) {
 					await MidiQOL.socket().executeAsGM("removeEffects", { actorUuid: targetToken.actor.uuid, effects: [effect.id] });
 				}
@@ -213,8 +178,9 @@ try {
 					content: `${targetToken.name}'s ${itemName} returns to normal.`,
 					speaker: ChatMessage.getSpeaker({ actor: actor })});
 			}
-			DAE.unsetFlag(actor, flagName);
 		}
+
+		await warpgate.revert(token.document, mutationName);
 	}
 	
 } catch (err) {