+++ scripts/spells/mass-cure-wounds.js
@@ -1,24 +1,23 @@
 /*
-	A wave of healing energy washes out from a point of your choice within range. Choose up to six creatures in a 30-foot-radius sphere centered on that point. Each target regains hit points equal to 3d8 + your spellcasting ability modifier. This spell has no effect on undead or constructs.
+	A wave of healing energy washes out from a point of your choice within range. Choose up to six creatures in a
+	30-foot-radius sphere centered on that point. Each target regains hit points equal to 3d8 + your spellcasting ability
+	modifier. This spell has no effect on undead or constructs.
 
-	At Higher Levels. When you cast this spell using a spell slot of 6th level or higher, the healing increases by 1d8 for each slot level above 5th.
+	At Higher Levels. When you cast this spell using a spell slot of 6th level or higher, the healing increases by 1d8
+	for each slot level above 5th.
 */
-const version = "10.0.0";
+const version = "11.0";
 const optionName = "Mass Cure Wounds";
 
 try {
-	const lastArg = args[args.length - 1];
-	const actor = MidiQOL.MQfromActorUuid(lastArg.actorUuid);
-	const actorToken = canvas.tokens.get(lastArg.tokenId);
-	
 	if (args[0].macroPass === "postActiveEffects") {
-		const level = Number(lastArg.spellLevel);
-		let targets = lastArg.targets;
+		const level = workflow.castData.castLevel;
+		let targets = workflow.targets;
 		let spellStat = actor.system.attributes.spellcasting;
 		if (spellStat === "") spellStat = "wis";
-		const spellcasting = actor.system.abilities[spellStat].mod;
+		const spellcastingModifier = actor.system.abilities[spellStat].mod;
 		
-		if (targets.length > 0) {
+		if (targets.size > 0) {
 			// build the target data
 			let choices = [];
 			let rows = "";
@@ -30,7 +29,7 @@ try {
 				if (!["construct", "undead"].includes(t.actor.system.details?.type?.value) && (totalHP < maxHP)) {
 					let row = `<div><input type="checkbox" style="margin-right:10px;"/><label>${t.actor.name}</label></div>`;
 					rows += row;
-					choices.push(t.object);
+					choices.push(t);
 				}
 			}
 			
@@ -67,9 +66,9 @@ try {
 							if (recipients.length > 0) {
 								let diceCount = 3 + level - 5;
 								
-								const healRoll = await new Roll(`${diceCount}d8 + ${spellcasting}`).evaluate({ async: false });
+								const healRoll = await new Roll(`${diceCount}d8 + ${spellcastingModifier}`).evaluate({ async: false });
 								await game.dice3d?.showForRoll(healRoll);
-								await new MidiQOL.DamageOnlyWorkflow(actor, actorToken, healRoll.total, "healing", recipients, healRoll, 
+								await new MidiQOL.DamageOnlyWorkflow(actor, token, healRoll.total, "healing", recipients, healRoll, 
 									{flavor: `${optionName}`, itemCardId: args[0].itemCardId});
 							}
 						}